requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
django-bootstrap3==14.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==1.7.0
PyJWT==1.7.1
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
zipp==3.4.0
